refactor: refactored seedeing from ufs for automated tests
