+++ src/modules/uf/seeders/uf.seeder.ts
@@ -1,6 +1,7 @@
 import { Seeder, SeederFactoryManager } from 'typeorm-extension';
 import { DataSource } from 'typeorm';
 import { Uf } from '../entity/uf.entity';
+import ufs from './uf.sample';
 
 export default class UfSeeder implements Seeder {
   public async run(
@@ -8,51 +9,6 @@ export default class UfSeeder implements Seeder {
     factoryManager: SeederFactoryManager,
   ): Promise<any> {
     const repository = dataSource.getRepository(Uf);
-    const ufs = [
-      {
-        name: 'Acre',
-        acronym: 'AC',
-      },
-
-      {
-        name: 'Alagoas',
-        acronym: 'AL',
-      },
-
-      {
-        name: 'Amazonas',
-        acronym: 'AM',
-      },
-
-      {
-        name: 'Amapá',
-        acronym: 'AP',
-      },
-
-      { name: 'Bahia', acronym: 'BA' },
-      { name: 'Ceará', acronym: 'CE' },
-      { name: 'Distrito Federal', acronym: 'DF' },
-      { name: 'Espírito Santo', acronym: 'ES' },
-      { name: 'Goiás', acronym: 'GO' },
-      { name: 'Maranhão', acronym: 'MA' },
-      { name: 'Mato Grosso', acronym: 'MT' },
-      { name: 'Mato Grosso do Sul', acronym: 'MS' },
-      { name: 'Minas Gerais', acronym: 'MG' },
-      { name: 'Pará', acronym: 'PA' },
-      { name: 'Paraíba', acronym: 'PB' },
-      { name: 'Paraná', acronym: 'PR' },
-      { name: 'Pernambuco', acronym: 'PE' },
-      { name: 'Piauí', acronym: 'PI' },
-      { name: 'Rio de Janeiro', acronym: 'RJ' },
-      { name: 'Rio Grande do Norte', acronym: 'RN' },
-      { name: 'Rio Grande do Sul', acronym: 'RS' },
-      { name: 'Rondônia', acronym: 'RO' },
-      { name: 'Roraima', acronym: 'RR' },
-      { name: 'Santa Catarina', acronym: 'SC' },
-      { name: 'São Paulo', acronym: 'SP' },
-      { name: 'Sergipe', acronym: 'SE' },
-      { name: 'Tocantins', acronym: 'TO' },
-    ];
     // ---------------------------------------------------
 
     const newUfs = await repository.create(ufs);
